Backend Product id verification

+++ backend/controllers/product.controller.js
@@ -35,6 +35,12 @@ export const updateProduct = async (req, res) => {
   const { id } = req.params;
   console.log("id:", id);
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -52,7 +58,7 @@ export const deleteProduct = async (req, res) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res
       .status(404)
-      .json({ success: false, message: "Inavlid product Id" });
+      .json({ success: false, message: "Invalid Product Id" });
   }
 
   try {
